Reverted previous examples/cgi_jpeg.cpp change.  It just covered up further embedded null handling problems in the library.  Fixed these by removing every last use of c_str() in lib/manip.cpp, using C++ string length information wherever possible in there, and changed const_string class to make copies of its data.
The latter is critical: when using SSQLS with BLOBs, the data gets
copied as ColData_Tmpl<const_string>, which used to just keep a
copy of the string pointer.  This didn't work out right with SSQLS,
because the pointed-to buffer went away after SSQLS population, so
the program could crash on trying to use the BLOB field.  (It didn't
always, which is why 2.3.0 got released with this bug.  Trying the
feature with a bigger BLOB tickled the bug.)


git-svn-id: svn://svn.gna.org/svn/mysqlpp/trunk@1651 5946f024-35f4-0310-b1b3-9a3be3380cfb