Broke coupling between the lifetime of a ResUse/Result object and the Row objects it creates.  It's now possible to reuse the result object while Row objects it created are still in use.  The only substantial reason for this were things having to do with looking up fields by name: we didn't want to copy the field name list into each Row object, so we had to keep the result object around to reference its copy.
Did we break down and make the wasteful copies?  No, we created
RefCountedPointer<T>, which lets us add reference counting to any object
(like FieldNames) and pass around copies of that object without
duplicating the underlying object unnecessarily.

This template will be used more in the near future.  This is the
beginning of the end for all the return-by-value of large objects
in MySQL++.


git-svn-id: svn://svn.gna.org/svn/mysqlpp/trunk@1751 5946f024-35f4-0310-b1b3-9a3be3380cfb